Add files via upload
It also allows you to select the audio files to be analyzed. You select your files easily and without the hassle of entering code in the terminal. It is analyzed. File Explorer runs again for you to choose where to save the output files. You choose the save location and it saves the output files there.